refactor(service): the assembly is engine-neutral, so it stops living behind an engine
