+++ src/pi.ts
@@ -60,8 +60,4 @@ export type { Model } from "@earendil-works/pi-ai";
 
 // The product's one-call assembly: a directory becomes a live service. On the pi surface because
 // it opens a directory the pi way (channels, schedules, the assembled agent).
-export {
-  createAgentService,
-  type AgentService,
-  type CreateAgentServiceOptions,
-} from "./service.ts";
+export { createAgentService, type CreateAgentServiceOptions } from "./engines/pi/service.ts";
